Make coredns embeddable.
